Feedback Final

+++ Feedback/AdminPage.php
@@ -8,84 +8,90 @@ if($_SESSION['admin'] == 0){
 if(isset($_POST['upload_fb'])){
     $query = "SELECT DISTINCT teacher_id,sub_id FROM feedback_temp;";
     $results = mysqli_query($db, $query);
-    while($row = mysqli_fetch_assoc($results)){
-        $t = $row['teacher_id'];
-        $s = $row['sub_id'];
-        $ans = array(0,0,0,0,0,0,0,0,0,0,0,0);
-        $rmrks = '';
-        $q = "SELECT * FROM feedback_temp WHERE teacher_id = '$t' AND sub_id = '$s';";
-        $res = mysqli_query($db, $q);
-        $length = mysqli_num_rows($res);
-        $cnt = array(array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0));
-        while($row1 = mysqli_fetch_assoc($res)){
-            $ans[0] = $ans[0] + (int)$row1['ques1'];
-            $cnt[0][(int)$row1['ques1'] - 1] = $cnt[0][(int)$row1['ques1'] - 1] + (int)1;
+    if(mysqli_num_rows($results) > 0){
+      while($row = mysqli_fetch_assoc($results)){
+          $t = $row['teacher_id'];
+          $s = $row['sub_id'];
+          $ans = array(0,0,0,0,0,0,0,0,0,0,0,0);
+          $rmrks = '';
+          $q = "SELECT * FROM feedback_temp WHERE teacher_id = '$t' AND sub_id = '$s';";
+          $res = mysqli_query($db, $q);
+          $length = mysqli_num_rows($res);
+          $cnt = array(array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0),array(0,0,0,0,0));
+          while($row1 = mysqli_fetch_assoc($res)){
+              $ans[0] = $ans[0] + (int)$row1['ques1'];
+              $cnt[0][(int)$row1['ques1'] - 1] = $cnt[0][(int)$row1['ques1'] - 1] + (int)1;
+              
+              $ans[1] = $ans[1] + (int)$row1['ques2'];
+              $cnt[1][(int)$row1['ques2'] - 1] = $cnt[1][(int)$row1['ques2'] - 1] + (int)1;
+              
+              $ans[2] = $ans[2] + (int)$row1['ques3'];
+              $cnt[2][(int)$row1['ques3'] - 1] = $cnt[2][(int)$row1['ques3'] - 1] + (int)1;
+              
+              $ans[3] = $ans[3] + (int)$row1['ques4'];
+              $cnt[3][(int)$row1['ques4'] - 1] = $cnt[3][(int)$row1['ques4'] - 1] + (int)1;
+              
+              $ans[4] = $ans[4] + (int)$row1['ques5'];
+              $cnt[4][(int)$row1['ques5'] - 1] = $cnt[4][(int)$row1['ques5'] - 1] + (int)1;
+              
+              $ans[5] = $ans[5] + (int)$row1['ques6'];
+              $cnt[5][(int)$row1['ques6'] - 1] = $cnt[5][(int)$row1['ques6'] - 1] + (int)1;
+              
+              $ans[6] = $ans[6] + (int)$row1['ques7'];
+              $cnt[6][(int)$row1['ques7'] - 1] = $cnt[6][(int)$row1['ques7'] - 1] + (int)1;
+              
+              $ans[7] = $ans[7] + (int)$row1['ques8'];
+              $cnt[7][(int)$row1['ques8'] - 1] = $cnt[7][(int)$row1['ques8'] - 1] + (int)1;
+              
+              $ans[8] = $ans[8] + (int)$row1['ques9'];
+              $cnt[8][(int)$row1['ques9'] - 1] = $cnt[8][(int)$row1['ques9'] - 1] + (int)1;
+              
+              $ans[9] = $ans[9] + (int)$row1['ques10'];
+              $cnt[9][(int)$row1['ques10'] - 1] = $cnt[9][(int)$row1['ques10'] - 1] + (int)1;
+              
+              $ans[10] = $ans[10] + (int)$row1['ques11'];
+              $cnt[10][(int)$row1['ques11'] - 1] = $cnt[10][(int)$row1['ques11'] - 1] + (int)1;
+              
+              $ans[11] = $ans[11] + (int)$row1['ques12'];
+              $cnt[11][(int)$row1['ques12'] - 1] = $cnt[11][(int)$row1['ques12'] - 1] + (int)1;
+              
+              $rmrks = $rmrks.'<->'.$row1['remark'];
+          }
+          // echo $t."-------".$s."<br><br>";
+          for($i = 0; $i < 12; $i++)
+          {
+            $c1 = $cnt[$i][0];
+            $c2 = $cnt[$i][1];
+            $c3 = $cnt[$i][2];
+            $c4 = $cnt[$i][3];
+            $c5 = $cnt[$i][4];
+            $qn = $i+1;
             
-            $ans[1] = $ans[1] + (int)$row1['ques2'];
-            $cnt[1][(int)$row1['ques2'] - 1] = $cnt[1][(int)$row1['ques2'] - 1] + (int)1;
-            
-            $ans[2] = $ans[2] + (int)$row1['ques3'];
-            $cnt[2][(int)$row1['ques3'] - 1] = $cnt[2][(int)$row1['ques3'] - 1] + (int)1;
-            
-            $ans[3] = $ans[3] + (int)$row1['ques4'];
-            $cnt[3][(int)$row1['ques4'] - 1] = $cnt[3][(int)$row1['ques4'] - 1] + (int)1;
-            
-            $ans[4] = $ans[4] + (int)$row1['ques5'];
-            $cnt[4][(int)$row1['ques5'] - 1] = $cnt[4][(int)$row1['ques5'] - 1] + (int)1;
-            
-            $ans[5] = $ans[5] + (int)$row1['ques6'];
-            $cnt[5][(int)$row1['ques6'] - 1] = $cnt[5][(int)$row1['ques6'] - 1] + (int)1;
-            
-            $ans[6] = $ans[6] + (int)$row1['ques7'];
-            $cnt[6][(int)$row1['ques7'] - 1] = $cnt[6][(int)$row1['ques7'] - 1] + (int)1;
-            
-            $ans[7] = $ans[7] + (int)$row1['ques8'];
-            $cnt[7][(int)$row1['ques8'] - 1] = $cnt[7][(int)$row1['ques8'] - 1] + (int)1;
-            
-            $ans[8] = $ans[8] + (int)$row1['ques9'];
-            $cnt[8][(int)$row1['ques9'] - 1] = $cnt[8][(int)$row1['ques9'] - 1] + (int)1;
-            
-            $ans[9] = $ans[9] + (int)$row1['ques10'];
-            $cnt[9][(int)$row1['ques10'] - 1] = $cnt[9][(int)$row1['ques10'] - 1] + (int)1;
-            
-            $ans[10] = $ans[10] + (int)$row1['ques11'];
-            $cnt[10][(int)$row1['ques11'] - 1] = $cnt[10][(int)$row1['ques11'] - 1] + (int)1;
-            
-            $ans[11] = $ans[11] + (int)$row1['ques12'];
-            $cnt[11][(int)$row1['ques12'] - 1] = $cnt[11][(int)$row1['ques12'] - 1] + (int)1;
-            
-            $rmrks = $rmrks.'<->'.$row1['remark'];
-        }
-        // echo $t."-------".$s."<br><br>";
-        for($i = 0; $i < 12; $i++)
-        {
-          $c1 = $cnt[$i][0];
-          $c2 = $cnt[$i][1];
-          $c3 = $cnt[$i][2];
-          $c4 = $cnt[$i][3];
-          $c5 = $cnt[$i][4];
-          $qn = $i+1;
+            // echo $c1."--".$c2."--".$c3."--".$c4."--".$c5."<br>";
+            $q = "INSERT INTO feedback_count(teacher_id, sub_id, question_no, count_1, count_2, count_3, count_4, count_5) VALUES ('$t','$s','$qn','$c1','$c2','$c3','$c4','$c5')";
+            mysqli_query($db,$q);
           
-          // echo $c1."--".$c2."--".$c3."--".$c4."--".$c5."<br>";
-          $q = "INSERT INTO feedback_count(teacher_id, sub_id, question_no, count_1, count_2, count_3, count_4, count_5) VALUES ('$t','$s','$qn','$c1','$c2','$c3','$c4','$c5')";
-          mysqli_query($db,$q);
-          
-        }
+          }
       
         
-        $avg_ans = array();
-        for($i=0;$i<12;$i++){
-            $avg_ans[$i] = strval(round((float)$ans[$i]/(float)$length));
-        }
-        $yr = strval(date('Y'));
-        $qu = "INSERT INTO feedback(teacher_id, sub_id, year, score1, score2, score3, score4, score5, score6, score7, score8, score9, score10, score11, score12, remark)";
-        $qu = $qu."VALUES ('$t','$s','$yr','$avg_ans[0]','$avg_ans[1]','$avg_ans[2]','$avg_ans[3]','$avg_ans[4]','$avg_ans[5]','$avg_ans[6]','$avg_ans[7]','$avg_ans[8]','$avg_ans[9]','$avg_ans[10]','$avg_ans[11]','$rmrks')";
-        mysqli_query($db, $qu);
-        
+          $avg_ans = array();
+          for($i=0;$i<12;$i++){
+              $avg_ans[$i] = strval(round((float)$ans[$i]/(float)$length));
+          }
+          $yr = strval(date('Y'));
+          $qu = "INSERT INTO feedback(teacher_id, sub_id, year, score1, score2, score3, score4, score5, score6, score7, score8, score9, score10, score11, score12, remark)";
+          $qu = $qu."VALUES ('$t','$s','$yr','$avg_ans[0]','$avg_ans[1]','$avg_ans[2]','$avg_ans[3]','$avg_ans[4]','$avg_ans[5]','$avg_ans[6]','$avg_ans[7]','$avg_ans[8]','$avg_ans[9]','$avg_ans[10]','$avg_ans[11]','$rmrks')";
+          mysqli_query($db, $qu);
+          
+      }
+      $qt = "TRUNCATE table feedback_temp";
+      mysqli_query($db, $qt);
+      echo "<script>alert('Data Uploaded!')</script>";
     }
-    
-    $qt = "TRUNCATE table feedback_temp";
-    mysqli_query($db, $qt);
+    else{
+      echo "<script>alert('No Data to Upload!')</script>";
+    }
+      
 
 }
 
@@ -201,7 +207,7 @@ function createpdf($fbresults, $teacherresults, $subjectresults,$subjectid,$teac
         $pdf->MultiCell(0, 10, ucfirst(strtolower($remarks[$x])), 0,'L');
         $x = $x-1;
     }
-    $pdf->Output('FeedBackOutputs/'.$teacherresults['teacher_dept'].'-'.$teacherresults['teacher_name'].'-'.$subjectresults['sub_name'].'-'.$fbresults['year'].'.pdf', 'F');
+    $pdf->Output('FeedbackOutputs/'.$teacherresults['teacher_dept'].'-'.$teacherresults['teacher_name'].'-'.$subjectresults['sub_name'].'-'.$fbresults['year'].'.pdf', 'F');
 
 }
 
@@ -211,24 +217,32 @@ if(isset($_POST['summary_fb'])){
 
     $q1 = "SELECT * FROM feedback";
     $result = mysqli_query($db, $q1);
+    
+    if(mysqli_num_rows($result)>0){
+      while($fbresults = mysqli_fetch_assoc($result)){
+          $subjectid = $fbresults['sub_id'];
+          $teacherid = $fbresults['teacher_id'];
 
-    while($fbresults = mysqli_fetch_assoc($result)){
-        $subjectid = $fbresults['sub_id'];
-        $teacherid = $fbresults['teacher_id'];
+          $q2 = "SELECT * FROM teacher WHERE teacher_id = '$teacherid'";
+          $teacherresults = mysqli_query($db, $q2);
+          $teacherresults = mysqli_fetch_array($teacherresults);
 
-        $q2 = "SELECT * FROM teacher WHERE teacher_id = '$teacherid'";
-        $teacherresults = mysqli_query($db, $q2);
-        $teacherresults = mysqli_fetch_array($teacherresults);
+          $q3 = "SELECT * FROM subject WHERE sub_id = '$subjectid'";
+          $subjectresults = mysqli_query($db, $q3);
+          $subjectresults = mysqli_fetch_array($subjectresults);
 
-        $q3 = "SELECT * FROM subject WHERE sub_id = '$subjectid'";
-        $subjectresults = mysqli_query($db, $q3);
-        $subjectresults = mysqli_fetch_array($subjectresults);
-
-        createpdf($fbresults, $teacherresults, $subjectresults,$subjectid,$teacherid);
-        
+          createpdf($fbresults, $teacherresults, $subjectresults,$subjectid,$teacherid);
+          
+      }
+      $qtr = "TRUNCATE table feedback_count";
+      mysqli_query($db, $qtr);
+      echo "<script>alert('Summary Created!')</script>";
     }
-    $qtr = "TRUNCATE table feedback_count";
-    mysqli_query($db, $qtr);
+    else{
+      echo "<script>alert('No data to summarize!')</script>";
+    }
+    
+    
 }
 if(isset($_POST['trunc_fb'])){
 
@@ -236,7 +250,7 @@ if(isset($_POST['trunc_fb'])){
     mysqli_query($db, $q1);
     $qt = "TRUNCATE table feedback_count";
     mysqli_query($db, $qt);
-    
+    echo "<script>alert('Data Cleared!')</script>";
 }
 
 ?>
